Add support for fish descriptions
Fish supports "descriptions" for completions, which are displayed to the
right of the value offered for completion, to provide additional context
for the user.

To display a description for a completion, the description is printed
after a tab character in the completion line.

so, instead of just printing:
```
bar
baz
qux
```

for completions, we can print:

```
bar<TAB>The bar command
baz<TAB>The baz command
qux<TAB>The qux command
```

This commit adds support for fish descriptions to tabry.

Instead of OptionsResults referencing a collection of `String`s, it now
holds a collection of `OptionResult`s, which hold the value of the
completion and an optional description.

Tabry will print a description if it is present, and if the
`TABRY_PRINT_DESCRIPTIONS` environment variable is set.