Adding withTabry nix helper
This commit adds a nix helper for binding tabry completion files with
nix packages. The `withTabry` function is a nix function that takes a
tabry completion file and a nix package and returns a derivation that
installs the completion file in the package's completion directory.

example usage:

```nix
wrappedCommand = withTabry ./command.tabry <pkg>
```